fun_bot/requirements.txt
# FunBot requirements (keep minimal/low-cost)
discord.py>=2.4.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0

# For Discord.py voice features
PyNaCl>=1.5.0
